requirements.txt
numpy
torchdiffeq
matplotlib
tqdm
scikit-learn
torchdiffeq
torchdyn